pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[dependency-groups]
dev = [
  "coverage[toml]>=6.4.4,<7",
  "factory-boy>=3.3.0,<4",
  "faker",
  "fixit>=2.2.1",
  "import-linter>=2.7",
  "mypy",
  "polyfactory>=3.3.0",
  "pre-commit",
  "pytest",
  "pytest-async-sqlalchemy",
  "pytest-asyncio",
  "pytest-cov",
  "pytest-factoryboy>=2.7.0,<3",
  "pytest-mock>=3.12.0,<4",
  "pytest-randomly",
  "ruff>=0.14.6"
]

[project]
authors = [{name = "jplhanna", email = "jpl.hanna@gmail.com"}]
dependencies = [
  "SQLAlchemy[asyncio]>=2.0.25,<3",
  "alembic~=1.8",
  "asyncpg>=0.30.0,<0.32",
  "dependency-injector>=4.46.0,<5",
  "discord.py>=2.0.1,<3",
  "furl>=2.1.3,<3",
  "psycopg2-binary>=2.9.6,<3",
  "pydantic>=2.5.3,<3",
  "pydantic-settings>=2.1.0,<3",
  "sqlmodel>=0.0.24,<0.1",
  "typing-extensions>=4.9.0,<5"
]
description = "Discord bot for tracking and rewarding tasks"
name = "discord-quest-bot"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.1"

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
omit = [
  "alembic/*",
  "run.py",
  "config.py",
  "conftest.py",
  "constants.py"
]

[tool.fixit]
disable = ["fixit.rules"]
enable = [".linting_rules"]
root = true

[tool.hatch.build.targets.sdist]
include = ["local_server"]

[tool.hatch.build.targets.wheel]
include = ["local_server"]

[tool.importlinter]
exclude_type_checking_imports = true
include_external_packages = true
root_package = "src"

[[tool.importlinter.contracts]]
containers = ["src"]
exhaustive = true
layers = [
  "tests",
  "conftest",
  "factories",
  "run | model_hub",
  "bot",
  "containers",
  "quests | tavern",
  "services",
  "repositories",
  "models",
  "helpers",
  "exceptions",
  "config",
  "constants",
  "typeshed"
]
name = "Source layers"
type = "layers"

[[tool.importlinter.contracts]]
containers = ["src.bot"]
layers = [
  "commands",
  "controllers",
  "typeshed",
  "constants"
]
name = "Bot layers"
type = "layers"

[[tool.importlinter.contracts]]
containers = ["src.quests", "src.tavern"]
exhaustive = true
layers = [
  "services",
  "exceptions",
  "models"
]
name = "Source Submodule layers"
type = "layers"

[tool.mypy]
cache_fine_grained = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
exclude = [
  "test",
  "alembic",
  "factories",
  "rule"
]
local_partial_types = true
namespace_packages = true
plugins = ["sqlalchemy.ext.mypy.plugin", "pydantic.mypy"]
python_version = "3.12"
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
disallow_untyped_calls = false
ignore_missing_imports = true
module = "sqlalchemy.*"

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = "furl"

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = "factory.*"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
asyncio_mode = "auto"
filterwarnings = [
  "ignore::DeprecationWarning"
]
markers = [
  "integration: tests which connect to a db"
]

[tool.ruff]
exclude = ["alembic"]
line-length = 120
target-version = "py313"

[tool.ruff.analyze]
# Exclude type-checking-only imports from the graph
type-checking-imports = false

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
fixable = ["ALL"]
ignore = ["D1", "D205", "TC006"]
select = [
  "E",
  "F",
  "W",
  "C",
  "PT",
  "I",
  "S",
  "D",
  "RSE",
  "RET501",
  "RET502",
  "RET503",
  "RET504",
  "FBT",
  "ASYNC",
  "UP",
  "SIM",
  "TID",
  "TCH",
  "ARG",
  "RUF100"
]

[tool.ruff.lint.isort]
force-single-line = true
force-to-top = ["services"]
forced-separate = ["src.tests", "src.helpers"]
lines-between-types = 1

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"src/factories/*" = ["FBT001"]
"src/model_hub.py" = ["F401"]
"src/tests/*" = ["S"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"
